Added momentum to player jump and Attack types to enemy
* [PlayerAgent] Adjusted the jump mechanic to incorporate momentum.
* [EnemyAgent] Added 'Typeofattack' Enum.

The player can now jump higher, based on the player's momentum prior to jumping.

The enemy has 3 types of attack (so far): 1) Use a weapon at range (range can be anything from 0 upwards). 2) Grab the player when next to player or 3) Grab player from a distance. The grabbing mechanics represent the worm's Wrap around the player' and 'Tongue' attacks.